backend/requirements.txt
asgiref==3.4.1
astroid==2.6.5
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Faker==8.10.3
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pylint==2.9.6
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
uritemplate==3.0.1
urllib3==1.26.6
validate-docbr==1.8.2
wrapt==1.12.1
